Software/Daemon/requirements.txt
# AlphaPaint Daemon - Python Dependencies

# Serial communication
pyserial>=3.5

# YAML configuration
PyYAML>=6.0
